+++ tests/testthat/test-ct--plot.R
@@ -11,3 +11,13 @@ test_that("No error with variable = 'density'", {
 
   expect_no_error( plot(ct_obj, variable = "density", cluster_id = cluster_id) )
 })
+
+test_that("plot.ct snapshot is constant", {
+  skip_on_ci()
+
+  ct_obj <- get_layer(test_sim_small(), "ct")
+  cluster_id <- get_layer_output(ct_obj)$cluster_id[1]
+  plot_obj <- plot(ct_obj, cluster_id = cluster_id)
+
+  expect_snapshot(plot_obj)
+})